Module Main.Make

Parameters

module Opt : Tool.Opt with type t = opt

module Conf : C.T

module TestRepr : Sc_lib.Ctypes_bridge.Struct.REPR

Signature

module Tool : Sc_lib.Tool.ToolInit

include Sc_lib.Tool.GenericInitializer with type input_data = [ `C ] Sc_sys.File.TYPES.file list with type working_data = Ty.project_data

include Sc_lib.Tool.TomlSpec

type opt

Options for the tool

val tool_name : string

The name of the tool

val toml_schema : opt Sc_lib.Eztoml.schema

Tool options, retrieved from project configuration

val default_opt : opt

A default option value

module TestRepr : Sc_lib.Ctypes_bridge.Struct.REPR

type input_data = [ `C ] Sc_sys.File.TYPES.file list

type working_data = Ty.project_data

val initialize : input_data -> (Sc_lib.Tool.project_status * (working_data, string) Stdlib.result) Lwt.t

Initializes a project. An input project is either a raw project to prepare for other tools, or a project already prepared that a tool may want to customize.

val post_process : working_data -> unit Lwt.t

Post processes the project if necessary